Thread local safe disabilitation added. Automatic creation of missing objects (self healing polymorphism)
